% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempt.R
\name{attempt}
\alias{attempt}
\title{Repeatedly try to come up with a valid result from a function}
\usage{
attempt(mainFun, valFun, maxAttempts, ...)
}
\arguments{
\item{mainFun}{The main generative function}

\item{valFun}{The function used to validate the answer from mainFun(); valid 
answers will return TRUE}
}
\value{
A result from the mainFun() function
}
\description{
This function allows for repeated tries (up to maxAttempts) of a function mainFun
to come up with result that is valid based on valFun.  Specifically, it is for 
times when an invalid results would be unfavorable but not catastrophically bad.
If it cannot do obtain a valid result, it will return the result of the last 
(i.e., invalid) attempt.
}
\examples{
##Set the main function
mydraw <- function(...) { rnorm(1, ...) }
##The validation function
myval <- function(rv) { (rv > 0) }
##Try to draw a positive value from the normal distribution
attempt(mydraw, myval, 1000, mean = 1, sd = 10)

}
\seealso{
attemptCrit
}

