% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attemptCrit.R
\name{attemptCrit}
\alias{attemptCrit}
\title{Repeatedly try to come up with a valid result from a function, stopping if failing}
\usage{
attemptCrit(mainFun, valFun, maxAttempts, ...)
}
\arguments{
\item{mainFun}{The main generative function}

\item{valFun}{The function used to validate the answer from mainFun(); valid
answers will return TRUE}
}
\value{
A result from the mainFun() function
}
\description{
This function allows for repeated tries (up to maxAttempts) of a function mainFun
to come up with result that is valid based on valFun.  Specifically, it is for
times when an invalid results would be catastrophically bad.  It stops with an
error if it is unable to come up with a valid answer
}
\examples{
##Set the main function
mydraw <- function(...) { rnorm(1, ...) }
##The validation function
myval <- function(rv) { (rv > 0) }
##Try to draw a positive value from the normal distribution
##Should work
attemptCrit(mydraw, myval, 1000, mean = 1, sd = 10)
##Should fail
attemptCrit(mydraw, myval, 10, mean = -100, sd = 0.01)
}
\seealso{
attemptCrit
}

