% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_functions.R
\name{calc_logAR}
\alias{calc_logAR}
\title{Calculate log-acceptance ratio for coefficient vectors or matrices}
\usage{
calc_logAR(data, outcome, Xmat, coef, coef_star, prior_mean = rep(0,
  length(coef)), prior_Sigma = diag(length(coef)) * 100,
  prop_Sigma = diag(length(coef)))
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Character containing outcome model variable name (e.g., "Y")}

\item{Xmat}{Design matrix corresponding to coefficient vector or matrix}

\item{coef}{Current value of coefficient vector or matrix}

\item{coef_star}{Proposed value of coefficient vector or matrix}

\item{prior_mean}{Prior distribution mean vector}

\item{prior_Sigma}{Prior distribution variance-covariance matrix}

\item{prop_Sigma}{Proposal distribution variance-covariance matrix}
}
\value{
Scalar log-acceptance ratio
}
\description{
Calculate log-acceptance ratio for coefficient vectors or matrices
}

