% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcLL.R
\name{calcLL}
\alias{calcLL}
\title{Log-likelihood}
\usage{
calcLL(v, par)
}
\arguments{
\item{v}{String containing variable name}

\item{par}{Parameter vector}
}
\value{
The sum of the log-likelihood of the data, as evaluated at "par"
}
\description{
Function to evaluate the total log-likelihood at a certain parameter set
}

