% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbern.R
\name{dbern}
\alias{dbern}
\title{Density of Bernoulli random variables}
\usage{
dbern(x, prob, ...)
}
\arguments{
\item{x}{Observed values of the Bernoulli R.V.}

\item{prob}{Scalar or vector of probabilities}
}
\value{
A vector containing the densities (or log-densities) of the observed values
}
\description{
This wrapper for dbinom() returns the density of one or more Bernoulli random
variables. Obtain the log-densities by passing "log=TRUE".
}
\examples{
##Get density of realizations with all different probabilities
dbern(x = c(1,0,0,1), prob = c(0.5,0.2,0.3,0.4))
}

