% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFamilyType.R
\name{detectFamilyType}
\alias{detectFamilyType}
\title{Detect type of family}
\usage{
detectFamilyType(family)
}
\arguments{
\item{family}{Family object to be parsed}
}
\value{
Character vector of "glm" or "multinom"
}
\description{
Returns whether the family passed is belongs fo GLM or VGLM multinomial
}
\examples{
##GLM
detectFamilyType(binomial(link="logit"))

##VGLM
detectFamilyType(multinomial(parallel=TRUE))

}

