% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawParamFromMLE.R
\name{drawParamFromMLE}
\alias{drawParamFromMLE}
\title{Draw a parameter from the asymptotic distribution of its MLE}
\usage{
drawParamFromMLE(v)
}
\arguments{
\item{v}{The name of the variable, as a string}
}
\value{
A parameter vector drawn from the asymptotic distribution of the fitted MLE
(if it converged) or the multivariate normal prior (if it did not)
}
\description{
It relies on the fact that there is an existing list in the environment
called "Xmats" that has a sublist with the variable's name and that sublist
contains the design matrix for the regression model where "V" is the outcome.
}
\details{
It also assumes that there is a similarly structured list called "fam" where the
family of "v" is stored, and a list called "params" that stores decent starting
values for the GLM (to assist in convergence).

It updates the "props" list with the new proposal mean and variance based on the
asymptotic distribution of the MLE from the regression fit
}
\examples{
##Make the lists that need to already be in the environment
props = list(Y = list(mean = c(0,0), sigma = diag(c(1,1))))
params = list(Y = c(0,0))
fam = list(Y = binomial(link="logit"))
Xmats = list(Y = cbind(1, rep(0:1, each = 5)))

##Make the outcome variable
Y = rbinom(n = 10, size = 1, prob = 0.5)

##Draw a vector from the prior
drawParamFromPrior(v = "Y")
}
\seealso{
n.rmvnorm, drawParam
}

