% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawParamFromPrior.R
\name{drawParamFromPrior}
\alias{drawParamFromPrior}
\title{Draw a parameter set from the (MVN) prior distribution}
\usage{
drawParamFromPrior(v, priorlist = NULL)
}
\arguments{
\item{v}{The name of the variable, as a string}

\item{priorlist}{(Optional) List containing priors; if not specified, then the
"priors" list must already exist in the surrounding environment}
}
\value{
A parameter vector drawn from the prior distribution
}
\description{
This draws from a multivariate normal density
It relies on the fact that there is an existing list in the environment called 
"priors" that has a sublist with the variable's name and that sublist contains 
the mean and variance-covariance matrix of the parameter's prior distribution
}
\examples{
##Make the prior list
priors = list(Y = list(mean = c(0,0), sigma = diag(c(1,1))))

##Draw a vector from the prior
drawParamFromPrior(v = "Y")

}
\seealso{
n.rmvnorm
}

