% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFrequentistModel.R
\name{fitFrequentistModel}
\alias{fitFrequentistModel}
\title{Fit a frequentist GLM or VGLM}
\usage{
fitFrequentistModel(v)
}
\arguments{
\item{v}{The name of the variable to use as the outcome in the regression}
}
\description{
Given a variable string name v, this function fits a GLM or VGLM based on the
design matrix found in the Xmats list Xmats[[v]] and the family found in the 
fam list in the larger environment
}
\examples{
##Set up lists in the environment
Xmats    <- list(Y = cbind(1, rep(1:5, each = 3)))
fam      <- list(Y = binomial(link="logit"))
params   <- list(Y = c(0,0))

##Make outcome variable with regression coefficient vector c(0,0.2)
Y <- rbinom(n = 15, size = 1, prob = 1/(1+exp(-X \%*\% c(0,0.2))))

##Call the function
summary(fitFrequentistModel("Y"))
}

