% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fromBSAe.R
\name{fromBSAe}
\alias{fromBSAe}
\title{Extract list or element in list from BSAe environment}
\usage{
fromBSAe(name, v = NULL)
}
\arguments{
\item{v}{The name of the element that should be returned;
If v is not specified, the entire list will be returned}
}
\value{
The  element of the version of Xmats stored in the BSEe
environment
}
\description{
This function pulls an element "v" from the list "name" stored in the BSAe
environment.
}
\examples{
BSAe <- new.env()
BSAe$params = list(Y = "this is what I want to get")
fromBSAe("params", "Y")
}

