% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genVar.R
\name{genVar}
\alias{genVar}
\title{Generate a random variable based on GLM framework}
\usage{
genVar(fam, Xmat, par)
}
\arguments{
\item{fam}{The family of the outcome, as would be passed to the GLM fit 
(like "binomial(link="logit")")}

\item{Xmat}{A design X matrix}

\item{par}{The vector of regression coefficients}
}
\description{
This function generates a random variable if passed an X matrix, a link, 
and a parameter set. For now, only Bernoulli random variables are supported.
}
\examples{
genVar(fam = binomial(link="logit"), Xmat = cbind(1, rnorm(10)), par = c(-1,1))

}
\seealso{
rbern
}

