% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pU1.R
\name{get.pU1}
\alias{get.pU1}
\title{Return the posterior probability of U equalling 1 for each observation}
\usage{
get.pU1()
}
\value{
A length-n vector of posterior P(U = 1)
}
\description{
This function is used in the imputation of U.  It calculates P(U=1|everything)
from the posterior.  This formula depends on link functions, and requires the
lists Xmats, as well as the string variable case (mediator or not) to exist in the environment.
}
\seealso{
Like
}

