% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLinkfun.R
\name{getLinkfun}
\alias{getLinkfun}
\title{Extract link function from a family object}
\usage{
getLinkfun(family)
}
\arguments{
\item{family}{A family object}
}
\value{
A function with the appropriate link
}
\description{
Returns a function corresponding to the desired link
}
\examples{
getLinkfun(binomial(link="logit"))
getLinkfun(multinomial(parallel=FALSE))
}
\seealso{
detectFamilyType
}

