% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLinkinv.R
\name{getLinkinv}
\alias{getLinkinv}
\title{Extract inverse link function from a family object}
\usage{
getLinkinv(family)
}
\arguments{
\item{family}{A family object}
}
\value{
A function with the appropriate link-inverse
}
\description{
Returns a function corresponding to the desired inverse link
}
\examples{
getLinkinv(binomial(link="logit"))
getLinkinv(multinomial(parallel=FALSE))
}
\seealso{
detectFamilyType
}

