% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getParam.R
\name{getParam}
\alias{getParam}
\title{Extract list or element in list from Xmats in the BSAe environment}
\usage{
getParam(v = NULL)
}
\arguments{
\item{v}{The name of the element that should be returned;
If v is not specified, the entire Xmats list will be returned}
}
\value{
The element of the version of Xmats stored in the BSEe
environment
}
\description{
This function pulls an element "v" from the list "Xmats" stored in the BSAe
environment.
}

