% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPrior.R
\name{getPrior}
\alias{getPrior}
\title{Extract list or element in list from priors in the BSAe environment}
\usage{
getPrior(v = NULL)
}
\arguments{
\item{v}{The name of the element that should be returned;
If v is not specified, the entire priors list will be returned}
}
\value{
The element of the version of priors stored in the BSEe
environment
}
\description{
This function pulls an element "v" from the list "priors" stored in the BSAe
environment.
}

