% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProp.R
\name{getProp}
\alias{getProp}
\title{Extract list or element in list from props in the BSAe environment}
\usage{
getProp(v = NULL)
}
\arguments{
\item{v}{The name of the element that should be returned;
If v is not specified, the entire props list will be returned}
}
\value{
The element of the version of props stored in the BSEe
environment
}
\description{
This function pulls an element "v" from the list "props" stored in the BSAe
environment.
}

