% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_functions.R
\name{indep}
\alias{indep}
\title{Independence sampler updating with fixed distribution
Proposes from a MVN with mean prop_mean and
variance-covariance matrix prop_Sigma}
\usage{
indep(coef, outcome, data, Xmat, ll_func, dprior, prop_mean, prop_Sigma, block)
}
\arguments{
\item{coef}{Coefficient vector or matrix}

\item{outcome}{Character containing outcome variable name in data}

\item{data}{Data frame}

\item{Xmat}{Design matrix for regression}

\item{ll_func}{Function to calculate log-likelihood for model}

\item{dprior}{Function to evaluate prior density for coefficients}

\item{prop_mean}{Mean of proposal distribution}

\item{prop_Sigma}{Variance-covariance matrix for proposal distribution}

\item{block}{Vector or matrix of same dimension of coef, denoting with which
block the parameter should be updated}
}
\value{
List containing updated coefficient (coef) and vector or matrix of
whether updates were accepted
}
\description{
Supports split block updating
}

