% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflatePriors.R
\name{inflatePriors}
\alias{inflatePriors}
\title{Inflated variance of prior distributions by appriorriate prior taus}
\usage{
inflatePriors()
}
\description{
If a prior tau exists in the environment (tauPriorY, tauPriorU, or tauPriorM,
if doing the mediator case), then this adjusts the prior by the appriorriate
factor
}
\seealso{
inflateProps
}

