% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflateProps.R
\name{inflateProps}
\alias{inflateProps}
\title{Inflated variance of proposal distributions by appropriate proposal taus}
\usage{
inflateProps()
}
\description{
If a proposal tau exists in the environment (tauPropY, tauPropU, or tauPropM,
if doing the mediator case), then this adjusts the proposal by the appropriate
factor
}

