% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflateVariance.R
\name{inflateVariance}
\alias{inflateVariance}
\title{Multiply specified rows of a covariance matrix by an inflation factor}
\usage{
inflateVariance(V, inflation = 1, indices = 1:ncol(V))
}
\arguments{
\item{V}{Covariance matrix}

\item{inflation}{Multiplicative factor for inflation}

\item{indices}{Row/column indices for columns to inflate}
}
\value{
Matrix the same shape as V
}
\description{
Multiply specified rows of a covariance matrix by an inflation factor
}

