% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initializeBSAe.R
\name{initializeBSAe}
\alias{initializeBSAe}
\title{Initialize BSAe environment for medBSA functions}
\usage{
initializeBSAe(quietly = FALSE)
}
\arguments{
\item{quietly}{Should the user be notified if an environment called BSEe
already exists? Defaults to FALSE, meaning that notification is given}
}
\description{
This function starts a fresh environment called "BSAe" which is used in many
functions relating to the BSA for mediation.  If "quietly" is set to FALSE,
it lets the user know if an environment of the same name has been overwitten.
}
\examples{
initializeBSAe()
}

