% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertZsIntoMean.R
\name{insertZsIntoMean}
\alias{insertZsIntoMean}
\title{Insert zeros into a mean vector}
\usage{
insertZsIntoMean(vec, p)
}
\arguments{
\item{vec}{Existing vector.  Elements must be named.}

\item{p}{Number of zeros to insert}
}
\value{
A vector of p elements longer than the vector supplied
}
\description{
This function takes a vector with named elements and inserts the desired
number of zeros into the middle of it. The zeros will be inserted after the
last "Z#" and will be named "Z(# + 1)" through "Z(# + p)", where p is the
number of zeros to insert.
}
\examples{
##Make vector
myVec <- 1:10
names(myVec) <- c(LETTERS[1:4], paste0("Z", 1:3), LETTERS[7:9])

##Insert two zeros into the vector after the last of the Z's
insertZsIntoMean(myVec, 2)
}
\seealso{
insertZsIntoVcov
}

