% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBinary.R
\name{isBinary}
\alias{isBinary}
\title{Check whether a family indicates the data is binary (actually, binomial)}
\usage{
isBinary(family)
}
\arguments{
\item{family}{A family object}
}
\value{
Logical TRUE or FALSE
}
\description{
Returns TRUE if a the family is binomial and FALSE otherwise
}
\examples{
##Should be TRUE
isBinary(binomial(link="logit"))

##Should be FALSE
isBinary(multinomial(parallel=FALSE))

}

