% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isMultinomialLogit.R
\name{isMultinomialLogit}
\alias{isMultinomialLogit}
\title{Check whether the family is multinomial logit}
\usage{
isMultinomialLogit(family)
}
\arguments{
\item{family}{A family object}
}
\value{
Logical TRUE or FALSE
}
\description{
This function returns TRUE if the family is multinomial with a logit
link and FALSE otherwise
}
\examples{
##Should be TRUE
isMultinomialLogit(family = multinomial(parallel=FALSE))

##Should be FALSE
isMultinomialLogit(family = binomial(link="logit"))
}

