% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Like.R
\name{Like}
\alias{Like}
\title{Calculate the likelihood of a variable given its design matrix and parameter}
\usage{
Like(v, Xmat = NULL, par = NULL, log = FALSE)
}
\arguments{
\item{v}{String containing the variable name}

\item{Xmat}{Optional parameter to specify design matrix used in linear predictor
part of v}

\item{par}{Optional parameter to specify parameter vector used in linear predictor
part of v}

\item{log}{Should the log-density be returned? Defaults to FALSE}
}
\value{
Vector of likelihood contributions for v
}
\description{
This function calculates the likelihood or log-likelihood of a variable given a
vector of realizations.  If Xmat and/or par are not explicitly provided, the
function extracts these from lists in the environment (Xmats and params)
}

