% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikMultinom.R
\name{logLikMultinom}
\alias{logLikMultinom}
\title{Make a dummy variable matrix based on}
\usage{
logLikMultinom(v, par)
}
\arguments{
\item{v}{String containing variable name}

\item{par}{Parameter vector of regression coefficients}
}
\description{
This function returns the log-likelihood of multinomial data.  Relies on the
existence of Xmats[[v]] and fam[[v]] containing the design matrix and link
functions for v.
}

