% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logPost.R
\name{logPost}
\alias{logPost}
\title{Function to return the log-posterior for a parameter set}
\usage{
logPost(v, par)
}
\arguments{
\item{v}{String containing the variable of interest}

\item{par}{Parameter vector of regression coefficients}
}
\description{
This function relies on existing lists in the environment (priors, Xmats) to
calculate the log-posterior of those parameter coefficients.
}

