% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPostMultinom.R
\name{logPostMultinom}
\alias{logPostMultinom}
\title{Function to return the log-posterior for a parameter set}
\usage{
logPostMultinom(v, par)
}
\arguments{
\item{v}{String containing the variable of interest}

\item{par}{Parameter vector of regression coefficients}
}
\description{
This function relies on existing lists in the environment (priors, Xmats) to 
calculate the log-posterior of those parameter coefficients.
}

