% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logProp.R
\name{logProp}
\alias{logProp}
\title{Calculate log-density of (multivariate normal) proposal distribution}
\usage{
logProp(v, par)
}
\arguments{
\item{v}{String containing variable name}

\item{par}{Value of parameter vector at which to evaluate log-density}
}
\description{
Based on the props list existing in the environment, this function calculates
the log-density of the proposal distribution evaluated at par
}

