% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy.R
\name{make_dummy}
\alias{make_dummy}
\title{Make dummy variable matrix}
\usage{
make_dummy(vec)
}
\arguments{
\item{vec}{Vector containing factor}
}
\value{
A matrix of dummy variables
}
\description{
Reference level is always the first level
}

