% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Xmats.R
\name{make_Xmat}
\alias{make_Xmat}
\title{Make design matrix from formula and data frame}
\usage{
make_Xmat(formula, data, ...)
}
\arguments{
\item{formula}{R formula object}

\item{data}{Data frame containing all variables on RHS of formula}
}
\value{
Matrix with appropriate column labels
}
\description{
Make design matrix from formula and data frame
}

