% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Xmats.R
\name{make_Xmat_set}
\alias{make_Xmat_set}
\title{Make version of design matrix where some covariate is set to
some constant value}
\usage{
make_Xmat_set(formula, data, varname, val)
}
\arguments{
\item{formula}{R formula object for outcome model}

\item{data}{Data frame containing all variables on RHS of formula}

\item{varname}{Name of covariate to "set" to something}

\item{val}{Value covariate should be "set" to}
}
\value{
Matrix with appropriate column labels
}
\description{
Preserves dummy coding if variable to set is a factor
}

