% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeMultinomialDummy.R
\name{makeMultinomialDummy}
\alias{makeMultinomialDummy}
\title{Expand a vector of integer values to the dummy variable version}
\usage{
makeMultinomialDummy(vec, ref = 1, dropRef = FALSE, prefix = NULL)
}
\arguments{
\item{vec}{Vector to be turned into indicator matrix}

\item{ref}{Reference category; if none specified, defaults to the lowest level}

\item{dropRef}{Should the reference column be dropped? Defaults to FALSE}

\item{prefix}{Prefix for column labels in returned matrix}
}
\value{
A matrix of indicator values for each of the levels
}
\description{
This function takes an integer vector and turns.  The lowest number is assumed to be
the reference level and the indicator variable column for this level
will be dropped from the matrix if dropRef is set to TRUE.
}

