% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriors.R
\name{makePriors}
\alias{makePriors}
\title{Function to make prior list based on MLE data fits}
\usage{
makePriors(data, formulas)
}
\arguments{
\item{data}{Data frame for fitting models}

\item{formulas}{List of formulas for regression models}
}
\value{
List of prior means and variances
}
\description{
(For now, assumes logit link and categorical M)
}

