% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeVar.R
\name{makeVar}
\alias{makeVar}
\title{Make a GLM-type variable based on the existing list values}
\usage{
makeVar(v, dimVar = NULL)
}
\arguments{
\item{v}{The string name of the variable to be generated}

\item{dimVar}{Only needed for categorical variable generation; number of unique
values the variable can take on}
}
\description{
This wrapper for genVar() and assumes there are existing lists fam, Xmats, and
params with element "v" that contain the necessary information.
}
\examples{
##Lists that need to exist already
fam  = list(Y = binomial(link="logit"))
Xmats = list(Y = cbind(1, rnorm(10)))
params = list(Y = c(-1,1))
##Make the variable based on these specifications
makeVar("Y")
}
\seealso{
genVar
}

