% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeZ.R
\name{makeZ}
\alias{makeZ}
\title{Generate Z variables (confounders) for simulation}
\usage{
makeZ()
}
\value{
A vector (if Zp = 1) or matrix (Zp > 1) of Z variables
}
\description{
This function generates a dataset of size ss with the Z parameters located
in the params[[Z]] list.  The Z's are created successively, so the Z1 depends 
on  nothing, while Z2 can depend on Z1, and so on.  The dimension of Z is 
determined by Zp.  When Zp > 1, params[["Z"]] should be list of lists containing
parameter vectors.
}
\examples{
##Variables that need to exist in environment
ss = 10
Zp = 2
fam = list(Z = list(binomial(link="logit"), binomial(link="logit")))
params = list(Z = list( c(0), c(0,1) ) )
##Make a Z matrix
makeZ()

}
\seealso{
genVar
}

