% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{medBSA}
\alias{medBSA}
\alias{medBSA-package}
\title{medBSA: a package for Bayesian sensitivity analysis for mediation}
\description{
medBSA: a package for Bayesian sensitivity analysis for mediation
}

