% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_functions.R
\name{metrop}
\alias{metrop}
\title{Random walk Metropolis updating
Proposes from a MVN with mean centered at current value and
variance-covariance matrix prop_Sigma}
\usage{
metrop(coef, outcome, data, Xmat, ll_func, dprior, prop_Sigma, block = coef *
  0 + 1)
}
\arguments{
\item{coef}{Coefficient vector or matrix}

\item{outcome}{Character containing outcome variable name in data}

\item{data}{Data frame}

\item{Xmat}{Design matrix for regression}

\item{ll_func}{Function to calculate log-likelihood for model}

\item{dprior}{Function to evaluate prior density for coefficients}

\item{prop_Sigma}{Variance-covariance matrix for proposal distribution}

\item{block}{Vector or matrix of same dimension of coef, denoting with which
block the parameter should be updated. Defaults to update all parameters at
once (all block = 1).}
}
\value{
List containing updated coefficient (coef) and vector or matrix of
whether updates were accepted
}
\description{
Supports split block updating
}

