% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n.rmvnorm.R
\name{n.rmvnorm}
\alias{n.rmvnorm}
\title{An as.numeric() version of a mvtnorm::rmvnorm()}
\usage{
n.rmvnorm(...)
}
\description{
Useful for ensuring sampled vectors will conform in matrix multiplication
Passes all arguments directly to rmvnorm()
}
\examples{
##Draw a multivariate random vector
n.rmvnorm(n = 1, mean = 1:3)
}

