% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstTru.R
\name{plotEstTru}
\alias{plotEstTru}
\title{Plot estimated parameters against true parameters}
\usage{
plotEstTru(est, tru, paramname)
}
\arguments{
\item{est}{The vector of parameter estimates}

\item{tru}{The vector of true parameter values corresponding to the estimates}

\item{paramname}{The label (used only in the title of the plot)}
}
\description{
This function takes a vector of parameter estimates, a vector of true values,
and a parameter name and it makes an attractive plot to demonstrate unbiasedness
and identifiability (or lack thereof)
}
\examples{
##Generate some values
tru = 1:100
est = rnorm(100, mean = tru, sd = 5)
##Plot them against each other
plotEstTru(est, tru, paramname = "My Favorite Parameter")

}

