% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_functions.R
\name{propose_coef}
\alias{propose_coef}
\title{Propose new values for a coefficient vector or matrix from an MVN}
\usage{
propose_coef(coef, prop_mean = c(coef), prop_Sigma)
}
\arguments{
\item{coef}{Current value of coefficient vector or matrix}

\item{prop_mean}{Mean of MVM proposal distribution. Defaults to current
value of coefficient vector (i.e., random vector)}

\item{prop_Sigma}{Variance-covariance matrix of MVM proposal distribution}
}
\value{
Vector or matrix of same shape as coef
}
\description{
Propose new values for a coefficient vector or matrix from an MVN
}

