% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbern.R
\name{rbern}
\alias{rbern}
\title{Sample Bernoulli random variables}
\usage{
rbern(prob, n = length(prob), ...)
}
\arguments{
\item{prob}{Scalar or vector of probabilities}

\item{n}{The number of Bernoulli random variable, in case there is a common probability. 
Defaults to the length of prob}
}
\value{
A vector of Bernoulli(prob) realizations
}
\description{
Shortcut wrapper for rbinom()
}
\examples{
##Generate variables with all different probabilities
rbern(seq(0.05,0.95,0.05)) 

}

