% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renameColsRows.R
\name{renameColsRows}
\alias{renameColsRows}
\title{Rename columns and rows of a square matrix}
\usage{
renameColsRows(mat, wantName, haveName)
}
\arguments{
\item{mat}{Square matrix in need of having column and row renamed}

\item{wantName}{New name for column and row that will get changed}

\item{haveName}{Current column name for the row/column of interest}
}
\description{
Rename columns and rows of a square matrix
}

