% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderRowsCols.R
\name{reorderRowsCols}
\alias{reorderRowsCols}
\title{Reorder the rows and columns of a square matrix using column names}
\usage{
reorderRowsCols(mat, wantOrder)
}
\arguments{
\item{mat}{Matrix with original column and row orderings}

\item{wantOrder}{Character list containing the desired order of column names}
}
\description{
Returns a matrix with columns and rows switched so that the columns (and
analogous rows) are in the desired order
}

