% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBSA.R, R/sampler.R
\name{runBSA}
\alias{runBSA}
\title{Function to run the sensitivity analysis}
\usage{
runBSA(data, formulas, priors, props, state, blocks = NULL, niter = 100)

runBSA(data, formulas, priors, props, state, blocks = NULL, niter = 100)
}
\description{
Function to run the sensitivity analysis

Function to run the sensitivity analysis
}

