% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneProposal.R
\name{tuneProposal}
\alias{tuneProposal}
\title{Function to inflate/deflate proposal variance}
\usage{
tuneProposal(props, acc, blocks, iter, lookback, tuneEvery, stopAt)
}
\arguments{
\item{props}{List of proposal distributions}

\item{acc}{List of acceptances}

\item{blocks}{Blocking of parameter updating}

\item{iter}{Current (just-completed) iteration of the MCMC chains}

\item{lookback}{How many iterations to look back to calculate
the acceptance rate}

\item{tuneEvery}{How frequently to potentially modify proposals}

\item{stopAt}{Iteration at which point proposal is no longer
modified (should be at most the burn-in)}
}
\value{
List of updated proposal distributions
}
\description{
Increases variance if accepting too often
Decreases variance if accepting too rarely
}

