% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unDummy.R
\name{unDummy}
\alias{unDummy}
\title{Collapse matrix of dummy variables into a single vector-valued version}
\usage{
unDummy(mat)
}
\arguments{
\item{mat}{Matrix of dummy variables (may or may not contain reference indicator)}
}
\description{
Reference category will be 0.
}
\examples{
##Version where all rows sum to 1 (i.e., reference column has been left in)
dummy.mat1 <- rbind(c(0,0,0,1), c(0,0,1,0), c(1,0,0,0))
unDummy(dummy.mat1)

##Version where some rows are all zeros (i.e., reference column removed)
dummy.mat2 <- rbind(c(0,0,1), c(0,1,0), c(0,0,0))
unDummy(dummy.mat2)

}

