% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateXmats.R
\name{updateXmats}
\alias{updateXmats}
\title{Update design matrices in BSAe environment with new value of U}
\usage{
updateXmats()
}
\description{
After a new U vector is imputed, this updates any matrices stored in the
Xmats list to the new value of U (uses the fact that the design matrix columns
are labelled).
}
\seealso{
get.pU1
}

