% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMpar.R
\name{validateMpar}
\alias{validateMpar}
\title{Validate M parameter vector (M.par)}
\usage{
validateMpar(par)
}
\arguments{
\item{par}{Vector of candidate M.par}
}
\value{
FALSE if M is log-link and has fitted values above 1; TRUE otherwise
}
\description{
If M has a log-link, it checks whether fitted values are all below 1.  If M 
is a log-link (fam[["M"]][["link"]] in the environment is log), then the check
will be performed. If the design matrix list Xmats exists and has element M,
then it checks that Xmats[["M"]] %*% par is less than 0 (meaning no fitted 
values out of range).  Otherwise, it checks to make sure that if all covariates
were in the range of (-1,1), no fitted values would be outside the [0,1] range
}
\examples{
##Make environment variables
fam = list(M = binomial(link="log"))
Zp = 1
Z = rbinom(n = 10, size = 1, prob = 0.5)
A = rbinom(n = 10, size = 1, prob = 0.5)

##Candidate M.pars (par1 should return TRUE, par2 should be FALSE)
par1 = c(0.2, 0.1, 2)
par2 = c(0.2, 0.1, -2)

##First type of check, where Xmats[["M"]] does not exist
validateMpar(par1)
validateMpar(par2)

#Second type of check, where Xmats[["M"]] does exist
Xmats = list(M = cbind(1,Z,A))
validateMpar(par1)
validateMpar(par2)

}

