% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateARD.R
\name{calculateARD}
\alias{calculateARD}
\title{Calculate average residual disparity}
\usage{
calculateARD()
}
\value{
ARD The average residual disparity after intervening on M
}
\description{
This function calculates the average residual disparity.  If the treatment A
can be a cause, this is equivalent to the controlled direct effect,
marginalized over the distribution of the confounders as observed in the A = 1
population.  For this to be meaningful as a residual disparity measure, A = 1
should be the relatively disadvantaged population (i.e., the one we would
intervene on).  Function requires A, Y.par, U.par, M.par, fam, and Xmats to
be in the environment
}

