% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikBin.R
\name{logLikBin}
\alias{logLikBin}
\title{Obtain log-likelihood of binary variable}
\usage{
logLikBin(v, par, Xmat = NULL, summation = TRUE)
}
\arguments{
\item{v}{String containing variable name}

\item{par}{Parameter vector of regression coefficients}

\item{Xmat}{Optional design matrix for computing log-likelihood; if none is
specified, Xmats[[v]] is used}

\item{summation}{Whether the individual likelihood contributions should be summed;
defaults to TRUE}
}
\description{
This function returns the log-likelihood of binary data.  Relies on the
existence of Xmats[[v]] and fam[[v]] containing the design matrix and link
functions for v.
}

