% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikMultinom.R
\name{logLikMultinom}
\alias{logLikMultinom}
\title{Calculate log-likelihood of multinomial random variable}
\usage{
logLikMultinom(v, par, Xmat = NULL)
}
\arguments{
\item{v}{String containing variable name}

\item{par}{Parameter vector of regression coefficients}

\item{Xmat}{Optional design matrix for computing log-likelihood; if none is
specified, Xmats[[v]] is used}
}
\description{
This function returns the log-likelihood of multinomial data.  Relies on the
existence of Xmats[[v]] and fam[[v]] containing the design matrix and link
functions for v.
}

