% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBSA.R
\name{runBSA}
\alias{runBSA}
\title{Function to run the sensitivity analysis}
\usage{
runBSA(data, formulas, priors, props, state = NULL, blocks = NULL,
  niter = 100)
}
\description{
Function to run the sensitivity analysis
}

